+++ configx/config_disable_features.go
@@ -1,0 +1,12 @@
+//go:build noquic
+
+package configx
+
+// DNS-over-QUIC
+type DOQServer struct {
+}
+
+// FIXME: dot.Server was a config file, should not be change.
+func (doq *DOQServer) parse() error {
+	return nil
+}
